requirements.txt
# Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Babel==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Caching==2.3.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Date/Time
python-dateutil==2.8.2
pytz==2023.3.post1
Babel==2.17.0

# Security
PyJWT==2.8.0

# Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
